requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
neo4j==5.15.0

# AI/ML Libraries
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.3

# Google AI
google-generativeai==0.3.2

# Document Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
python-multipart==0.0.6

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0

# Text Processing
tiktoken==0.5.2
nltk==3.8.1

# Data Processing
pandas==2.1.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0